Automatic detection of metric interval
If a metric is received with an Interval field set to 0, then we
need to automatically detect the interval.  This is achieved by buffering
the last 3 points in memory.  We then sort the points by timestamp
and get the time difference between the first 2, and last 2.  The lower
difference is used as the interval. Metrics are delayed from being processed
until the Interval is known, so when a new series is sent it will take 3x interval
before any data is queriable.

Future effort needs to be done to adjust
the interval to known valid intervals. eg 1s, 5s, 10s, 30s, 1m, 5m, etc....